Use submission list from base ref for duplicate URL check
Previously, the duplicate URL check was configured to use the submission list at the head ref, where two ocurrences of
the URL indicated a duplicate. A more secure approach is to use the submission list at the base ref, and this means that
a single ocurrence of the URL indicates a duplicate.