Prevent linkification of example URL in warning message
When the submission is not a valid Git clone URL (e.g., they submitted the URL to a release page), a warning message is displayed:

Submission URL is not a Git clone URL (e.g., https://github.com/arduino-libraries/Servo).

GitHub automatically linkifies the URL in the message when it is presented as a comment in the PR thread. I feel this
linkification makes it less clear that the URL itself is intended to be the example. Wrapping it in backticks prevents
the linkification.