Update link shown when submission is on a supported Git host
When the submission URL is not from on one of the Git hosts currently supported by the indexer, an error message is
displayed which provides a link to the section of the Library Manager URL that lists the submission requirements, along
with an explanation that addition of new major hosts can be considered on request.

That link points to the Library Manager FAQ hosted in the arduino/Arduino repository's wiki. That content has since been
migrated to the new repository, and will eventually be replaced by a link to the new location once it is public. So it's
best to point the URL directly to the new location for this information.

Due to an organization of the FAQ content, the most relevant content is now under a different heading, thus the anchor ID
change.