Handle root submission URLs
Previously, a submission URL with a root path resulted in a panic:

panic: Rel: can't make .git relative to /

These URLs won't represent a valid submission, but panics should only occur when the error should not occur under any
normal circumstances. By avoiding the panic, the process will fail with the helpful error either that the URL is not
under a supported Git host (e.g., https://mycoollibrary.com as submission), or that it is not a valid Git clone URL
(e.g., https://github.com as submission).