Check for duplicates within the submission itself
A check for duplication between the submission and the libraries already on the list was already done. However, it is
possible that submitters will add the same library twice within a single submission. This was not previously checked for.
In addition to causing confusion at a later time for people working directly with the list, it's also possible that
allowing duplicates into the list could have harmful implications for the indexer or for Library Manager itself.

So it's safest to make a check for duplicates within the submission and return a helpful error message, requiring the
submission to contain only unique libraries before accepting it.