Output appropriate `arduino-lint --library-manager` setting
When a submission URL is modified, it must also be checked. In this case, the `arduino/arduino-lint-action`'s
`library-manager` input setting must be "update" instead of "submit", since the library is already in the index.

Since, manual testing of GitHub Actions workflows is not well supported, it's best to keep as much of the logic as
possible in the Go code of the parser. For this reason, the parser should determine the appropriate `library-manager`
setting and pass that to the workflow.