Configure Dependabot for Go and Python dependencies
Dependabot is already in use to maintain the dependencies of the project's GitHub Actions workflows. It will be helpful
to do the same with the Go dependencies.

With the introduction of Python-based integration tests, the Python dependencies of the project became more significant.
So we might as well use Dependabot for them too!