Prevent linkification of URL in error message
GitHub Flavored Markdown automatically linkifies URLs and web addresses. A link is a "click me" signal to the reader. In
this case, there is no reason for the reader to visit the web address. The error message is only displaying it to make it
clear what is being referred to. So the linkification actually has a harmful effect on readability and might sidetrack
the reader.