Report removal of final newline
If a file does not end in a newline, then every line added to the end of the file results in a diff to the previous final
line.

An example using `\n` to symbolize end of line characters:

https://github.com/arduino-libraries/Servo\n
https://github.com/arduino-libraries/Stepper

When another line is added at the end:

https://github.com/arduino-libraries/Servo\n
https://github.com/arduino-libraries/Stepper\n
https://github.com/arduino-libraries/FooBar

you can see there is necessarily a diff on line 2 even though the editor did not touch that line.

For this reason, it is best practices to always retain a newline at the end of files.

In addition to it being common practice, the GitHub web editor automatically adds this newline. Between this and my
expectation that people would either add URLs to the top of the list or in alphabetical order, leaving the distant end of
the file alone, I was hoping that this would not be a common problem with the library submission system, but this is
simply not a realistic expectation.

What I hadn’t considered is that the person to suffer for this formatting is the next to add a URL to the end of the
file. Because of this “ghost diff”, the submission system sees this PR as a modification to an existing item on the list
rather than an addition.

I think this will be best handled by detecting and dealing with the problem at its introduction, rather than attempting
to make the system resilient to a missing final newline.