Move Python dependencies to "dev" section of manifest
Even though the dev/main dichotomy is not very meaningful to this particular project, which only uses Python for
development infrastructure, the move toward standardization of "template" assets that can be applied to any project with
the minimum of customization means that the infrastructure might be applied to other projects which do have
non-development Python dependencies. Since there is no real benefit to putting development dependencies in the "main"
section, it is best to be consistent in the way these are handled across all projects.