Run build dep of integration test task in parallel with Python deps install
The tasks called from the `deps` key of a task are run in parallel. Once they are finished, the commands from the `cmds`
key are run in serial. In the case of tasks that don't depend on the output from another, this can make the task more
efficient.

Such is the case with the `go:test-integration` task. The Go application must be built and the Python testing
dependencies installed before the tests can run, but those two task dependencies can run in parallel.