Use standardized command in `go:build` task
This is the command used in the standardized "template" task that should be used in all Go projects.

The previous `GO_BUILD_FLAGS` variable was intended to allow the developer to specify arbitrary arguments via the command
line, but that is not needed.

The `LDFLAGS` variable that is introduced here allows injecting data to the build, which is used in more complex
applications, but not in this one at the moment.