Add support to infrastructure assets for multiple Go modules
The "templates" for testing, linting, and formatting Go code work from a list of packages of the Go module which is
output from a `go list` command. Previously, it was only possible to get the packages of the module in the root of the
repository.

Projects may contain multiple Go modules in subfolders of the repository. In order to support checks on these modules,
it's necessary to configure the commands to run from their path. This is passed to the task via the `GO_MODULE_PATH`
environment variable. If this variable is not defined, the default root module path is used as default, preserving the
previous task behavior.

The workflows use a job matrix to allow easy configuration for any number of module paths and a dedicated parallel job
for each module.

Because it may not be desirable to combine the code coverage data for all modules, a Codecov flag value can be defined
for each matrix job in the "Test Go" workflow. This workflow produces a two dimensional "operating system * module" job
matrix, running the tests for each module using all operating systems.

Even though this project has, and is likely to alway have, only the one module in the common location in the root of the
repo, and so was perfectly well served by the previous infrastructure, these assets are intended to be applicable to any
project, and so must be more extendable. The new approach works just as well as before for this project.