Add check for missed Go module tidying to "Check Go" workflow
It can sometimes be easy to forget to run a `go mod tidy` to sync up the modules dependencies in the manifest and lock
files. This often results in a confusing diff for the next contributor making a change that should not result in any
change to those files.

The better approach is to check the tidy state on every change to a relevant file so the module can be kept in a clean
state.