Add `LDFLAGS` taskfile variable
This taskfile variable is used to inject data into the build process. It is not used in this project, but the "template"
`go:build` task still uses this variable in its command.

Even though it makes no functional difference, I think the workflow is less confusing with an empty definition of the
variable to make it clear that this was intended, rather than being an accidentally lost declaration.