Correct argument order in test assertions
The `testify` Go packages use the "Yoda condition"-style argument order: expected, actual. If the assertions reverse
the order of the arguments, the test failure output becomes very confusing.

Even though `assert.ElementsMatch()` doesn't treat the order of the arguments specially, it is best to be consistent
with the order.