Use the "other" type for newline-only diffs
In order to make the system as lenient as possible in regards to submission format, newline changes to the list are
ignored (with the exception of final newline removal). For this reason, a PR consisting solely of changes to the newlines
in the list results in a null list of URLs, as would be expected.

The problem was that these PRs were assigned a "modification" type. A workflow job matrix is used to check the list of
URLs generated from a "modification" request "update" mode to see whether they are still compliant with the requirements
in order to triage modification requests before the registry maintainer does the final manual review of the PR.

GitHub Actions does not allow zero length job matrices, and anyway this sort of PR does not match with the meaning we
have assigned the "modification" type. The "other" type is most appropriate for this type of PR. "other" type requests
are also manually reviewed, but without the automated triage.