Adjust code for breaking change to github.com/arduino/go-paths-helper
A breaking change was made to the `github.com/arduino/go-paths-helper.Path.IsInsideDir` method's signature in the 1.9.0
release. The error result parameter was removed from the method because the conditions that produced the "error" state
may actually only be the result of the normal use case of comparing between paths on different file systems.

For this reason, the fix for the breaking change is to simply remove the error handling for the method call.