Use a dedicated output field to indicate request conclusion
In cases where the request submitter or the submitted repositories are from entities without registry access privileges,
the parser can determine the conclusion of the request. When the access control system was implemented, the pre-existing
`type` field was used to pass this information. That field is used in two ways:

- Determine the appropriate handling for the request.
- Apply the appropriate labeling to the pull request.

Although the `type` field is suitable for the first of these in the case where the request is declined, it is not very
suitable for the second. The reason is that the labels are grouped in sets. The type data corresponds to the "topic"
label set, while the conclusion corresponds to the "conclusion" set. The PR should be labeled with the appropriate label
from each of these sets.

So the request conclusion data is moved to a dedicated `conclusion` output field, leaving the `type` field to be used
exclusively for its original purpose.