Bump "Licensed" version in dependency license check workflow
The version of the "Licensed" tool for use in the GitHub Actions workflow is defined via the "licensee/setup-licensed"
action's `version` input.

Previously the action was configured to install version 3.x of the action. That version is significantly outdated. The
workflow is hereby updated to use the latest version of Licensed.