Show resolved URL in duplicate submission error message when relevant
library-registry-submission-parser resolves redirects in the submitted URLs and validates the request for that resolved
URL.

One of the validations is checking whether the submitted URL is already in the repositories list. If so, the validation
fails and the bot makes a comment on the pull request. In the case where the resolved URL is different from the
submitted URL, this failure can be confusing, since the duplicate will not be found in a search for the submitted URL.
Previously, the comment from the bot only mentioned the submitted URL.

The communication about the validation failure under these conditions is hereby improved by using a different error
message that explains it is the resolved URL that was found in the list, and states that URL.